Refactor endpoint rule execution helpers
